fixes from conversation threads
