+++ src/lib/server/job-executors/index.ts
@@ -6,4 +6,3 @@ export * as Publish from './publish';
 export * as System from './system';
 export * as UserTasks from './userTasks';
 export * as Polling from './polling';
-export * as UDMCleanup from './databaseCleanup';
