+++ src/lib/server/job-executors/system/index.ts
@@ -1,5 +1,6 @@
+import { cleanup } from './cleanup';
 import { checkPendingUpdates, checkSystemStatuses } from './engine';
 import { refreshLangTags } from './langtags';
 import { lazyMigrate, migrate } from './migrate';
 
-export { checkPendingUpdates, checkSystemStatuses, lazyMigrate, migrate, refreshLangTags };
+export { checkPendingUpdates, checkSystemStatuses, cleanup, lazyMigrate, migrate, refreshLangTags };
